cheri: Apply the getpage.cocci semantic patch
Calls to get_page variants that return an address instead of
a pointer but be treated as a uintptr_t not as unsigned long.
If the return value is assigned to a variable change its type
to uintptr_t.

Apply like this:

	make COCCI=scripts/coccinelle/getpage.cocci \
		MODE=patch coccicheck

+++ arch/um/kernel/skas/mmu.c
@@ -26,7 +26,7 @@ static struct list_head mm_list;
 int init_new_context(struct task_struct *task, struct mm_struct *mm)
 {
 	struct mm_id *new_id = &mm->context.id;
-	unsigned long stack = 0;
+	uintptr_t stack = 0;
 	int ret = -ENOMEM;
 
 	stack = __get_free_pages(GFP_KERNEL | __GFP_ZERO, ilog2(STUB_DATA_PAGES));
